fix: rp_att type is now float and not int64

+++ env/psp_env_specification.py
@@ -145,9 +145,7 @@ class PSPEnvSpecification:
                     shape=self.shape_rp,
                     dtype=np.int64,
                 ),
-                "rp_att": Box(
-                    low=0, high=1000, shape=self.shape_rp_att, dtype=np.int64
-                ),
+                "rp_att": Box(low=0, high=1000, shape=self.shape_rp_att, dtype=float),
             }
         )
